+++ thread_test_doctest.cpp
@@ -37,7 +37,7 @@ TEST_CASE("VdCorput thread safety") {
     }
 
     std::sort(all_values.begin(), all_values.end());
-    for (size_t i = 1; i < all_values.size(); ++i) {
+    for (int i = 1; i < all_values.size(); ++i) {
         CHECK_NE(all_values[i], all_values[i - 1]);
     }
